[TAN-5920] Refactored project import preview into a separate class
